# client/awast-front/src/assets/swagger/swagger.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/api/v1/zap/spider":{"post":{"tags":["scanner"],"summary":"Zap Spider","operationId":"zap_spider_api_v1_zap_spider_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestBody"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/zap/spider_status/{scan_id}":{"get":{"tags":["scanner"],"summary":"Zap Spider Status","operationId":"zap_spider_status_api_v1_zap_spider_status__scan_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"scan_id","in":"path","required":true,"schema":{"type":"string","title":"Scan Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/zap/scan":{"post":{"tags":["scanner"],"summary":"Zap Scan","operationId":"zap_scan_api_v1_zap_scan_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestBody"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/zap/scan_status/{scan_id}":{"get":{"tags":["scanner"],"summary":"Zap Scan Status","operationId":"zap_scan_status_api_v1_zap_scan_status__scan_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"scan_id","in":"path","required":true,"schema":{"type":"string","title":"Scan Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/zap/alerts":{"get":{"tags":["scanner"],"summary":"Zap Alerts","operationId":"zap_alerts_api_v1_zap_alerts_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/zap/alerts/target":{"get":{"tags":["scanner"],"summary":"Zap Alerts With Evidence","operationId":"zap_alerts_with_evidence_api_v1_zap_alerts_target_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"baseurl","in":"query","required":false,"schema":{"type":"string","title":"Baseurl"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/zap/alerts/summary":{"get":{"tags":["scanner"],"summary":"Zap Alerts Summary","operationId":"zap_alerts_summary_api_v1_zap_alerts_summary_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/zap/alerts/{alert_id}":{"get":{"tags":["scanner"],"summary":"Zap Alert By Id","operationId":"zap_alert_by_id_api_v1_zap_alerts__alert_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"alert_id","in":"path","required":true,"schema":{"type":"string","title":"Alert Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/zap/abort/scan/{scan_id}":{"get":{"tags":["scanner"],"summary":"Zap Abort","operationId":"zap_abort_api_v1_zap_abort_scan__scan_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"scan_id","in":"path","required":true,"schema":{"type":"string","title":"Scan Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/exploiter/run":{"post":{"tags":["exploiter"],"summary":"Exploiter Run","operationId":"exploiter_run_api_v1_exploiter_run_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExploiterRequestBody"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/report/new":{"post":{"tags":["report"],"summary":"Report New","operationId":"report_new_api_v1_report_new_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/report/download":{"post":{"tags":["report"],"summary":"Report Download","operationId":"report_download_api_v1_report_download_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DownloadReportRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/user/all":{"get":{"tags":["user"],"summary":"Get Users","operationId":"get_users_api_v1_user_all_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/UserOut"},"type":"array","title":"Response Get Users Api V1 User All Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/user/me":{"get":{"tags":["user"],"summary":"Read Profile","operationId":"read_profile_api_v1_user_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/user/{user_id}":{"get":{"tags":["user"],"summary":"Get User","operationId":"get_user_api_v1_user__user_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["user"],"summary":"Update User","operationId":"update_user_api_v1_user__user_id__patch","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["user"],"summary":"Delete User","operationId":"delete_user_api_v1_user__user_id__delete","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"},"title":"Response Delete User Api V1 User  User Id  Delete"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/user/new":{"post":{"tags":["user"],"summary":"Create User","operationId":"create_user_api_v1_user_new_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/user/my/scan/history":{"get":{"tags":["user"],"summary":"Get My Scan History","operationId":"get_my_scan_history_api_v1_user_my_scan_history_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/user/my/scan/{scan_id}/results":{"get":{"tags":["user"],"summary":"Get My Scan History","operationId":"get_my_scan_history_api_v1_user_my_scan__scan_id__results_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"scan_id","in":"path","required":true,"schema":{"type":"string","title":"Scan Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["auth"],"summary":"Login","operationId":"login_api_v1_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLogin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/refresh":{"post":{"tags":["auth"],"summary":"Refresh Token Route","operationId":"refresh_token_route_api_v1_auth_refresh_post","requestBody":{"content":{"application/json":{"schema":{"type":"string","title":"Refresh Token"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/analyze/swagger":{"post":{"summary":"Analyze Swagger File","operationId":"analyze_swagger_file_api_v1_analyze_swagger_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_analyze_swagger_file_api_v1_analyze_swagger_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SwaggerAnalysisResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_analyze_swagger_file_api_v1_analyze_swagger_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_analyze_swagger_file_api_v1_analyze_swagger_post"},"DownloadReportRequest":{"properties":{"report_id":{"type":"string","title":"Report Id"}},"type":"object","required":["report_id"],"title":"DownloadReportRequest"},"ExploiterRequestBody":{"properties":{"target":{"type":"string","title":"Target"},"params":{"type":"string","title":"Params"},"vuln_type":{"type":"string","title":"Vuln Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"login_url":{"type":"string","title":"Login Url"},"method":{"type":"string","title":"Method"}},"type":"object","required":["target","params","vuln_type","username","password","login_url","method"],"title":"ExploiterRequestBody"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ReportRequest":{"properties":{"scan_id":{"type":"string","title":"Scan Id"}},"type":"object","required":["scan_id"],"title":"ReportRequest"},"RequestBody":{"properties":{"target":{"type":"string","title":"Target"}},"type":"object","required":["target"],"title":"RequestBody"},"SwaggerAnalysisResponse":{"properties":{"findings":{"items":{"$ref":"#/components/schemas/VulnerabilityFinding"},"type":"array","title":"Findings"},"total_findings":{"type":"integer","title":"Total Findings"}},"type":"object","required":["findings","total_findings"],"title":"SwaggerAnalysisResponse"},"UserCreate":{"properties":{"username":{"type":"string","title":"Username"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"},"role":{"type":"string","title":"Role"},"enabled_domains":{"items":{"type":"string"},"type":"array","title":"Enabled Domains"}},"type":"object","required":["username","email","password","role","enabled_domains"],"title":"UserCreate"},"UserLogin":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"UserLogin"},"UserOut":{"properties":{"user_id":{"type":"string","format":"uuid","title":"User Id"},"username":{"type":"string","title":"Username"},"email":{"type":"string","format":"email","title":"Email"},"enabled_domains":{"items":{"type":"string"},"type":"array","title":"Enabled Domains"},"role":{"type":"string","title":"Role"}},"type":"object","required":["user_id","username","email","enabled_domains","role"],"title":"UserOut"},"UserUpdate":{"properties":{"username":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Username"},"email":{"anyOf":[{"type":"string","format":"email"},{"type":"null"}],"title":"Email"},"role":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Role"},"password":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Password"},"enabled_domains":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Enabled Domains"}},"type":"object","title":"UserUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VulnerabilityFinding":{"properties":{"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"severity":{"type":"string","title":"Severity"},"location":{"type":"string","title":"Location"},"remediation":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Remediation"}},"type":"object","required":["title","description","severity","location"],"title":"VulnerabilityFinding"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/auth/login"}}}}}}